+++ src/Products/PloneMeeting/safe_utils.py
@@ -38,6 +38,7 @@ from Products.PloneMeeting.esign.utils import get_meeting_esign_signatories
 from Products.PloneMeeting.ftw_labels.utils import get_labels
 from Products.PloneMeeting.utils import cleanMemoize
 from Products.PloneMeeting.utils import cropHTML
+from Products.PloneMeeting.utils import display_as_html
 from Products.PloneMeeting.utils import down_or_up_wf
 from Products.PloneMeeting.utils import escape
 from Products.PloneMeeting.utils import field_is_empty