+++ CHANGES.rst
@@ -15,6 +15,15 @@ Changelog
 - Reordered fields at end of item view in a more chronological way:
   `committeeObservations`, `committeeTranscript`, `inAndOutMoves` and `notes`.
   [gbastien]
+- Do not `saveHasActions` the `imio.actionspanel backlink` action as it is
+  not displayed as an icon or it prevent displaying `-` when no action available.
+  [gbastien]
+- Added searches `searchitemsofmygroupswithneededfollowup` and
+  `searchitemsofmygroupswithprovidedfollowup`.
+  [gbastien]
+- Added `utils.display_as_html` so it is available in expressions to render
+  `text/plain` as `text/html`.
+  [gbastien]
 - Display potential session title in prepare for signature forms.
   [gbastien]